backward-compatibilities with old clients

+++ ping.php
@@ -17,14 +17,6 @@
     if (!isset( $_REQUEST['players'] )) exit();
     if (!isset( $_REQUEST['mods'] )) exit();
     if (!isset( $_REQUEST['map'] )) exit();
-    if (!isset( $_REQUEST['maxplayers'] )) exit();
-    if (!isset( $_REQUEST['title'] )) exit();
-    if (!isset( $_REQUEST['description'] )) exit();
-    if (!isset( $_REQUEST['type'] )) exit();
-    if (!isset( $_REQUEST['width'] )) exit();
-    if (!isset( $_REQUEST['height'] )) exit();
-    if (!isset( $_REQUEST['tileset'] )) exit();
-    if (!isset( $_REQUEST['author'] )) exit();
     
     header( 'Content-type: text/plain' );
     try 
@@ -46,6 +38,79 @@
                 $state = -1;
 		}
 		
+		// required support of old clients
+		if ( isset($_REQUEST['maxplayers']))
+		{
+			$maxplayers = $_REQUEST['maxplayers'];
+		}
+		else
+		{
+			$maxplayers = 0;
+		}
+		
+		if ( isset($_REQUEST['title']))
+		{
+			$title = $_REQUEST['title'];
+		}
+		else
+		{
+			$title = '';
+		}
+		
+		if ( isset($_REQUEST['description']))
+		{
+			$description = $_REQUEST['description'];
+		}
+		else
+		{
+			$description = '';
+		}
+		
+		if ( isset($_REQUEST['type']))
+		{
+			$type = $_REQUEST['type'];
+		}
+		else
+		{
+			$type = '';
+		}
+		
+		if ( isset($_REQUEST['width']))
+		{
+			$width = $_REQUEST['width'];
+		}
+		else
+		{
+			$width = 0;
+		}
+		
+		if ( isset($_REQUEST['height']))
+		{
+			$height = $_REQUEST['height'];
+		}
+		else
+		{
+			$height = 0;
+		}
+		
+		if ( isset($_REQUEST['tileset']))
+		{
+			$tileset = $_REQUEST['tileset'];
+		}
+		else
+		{
+			$tileset = '';
+		}
+		
+		if ( isset($_REQUEST['author']))
+		{
+			$author = $_REQUEST['author'];
+		}
+		else
+		{
+			$author = '';
+		}
+		
         $insert = $db->prepare('INSERT OR REPLACE INTO servers 
             (name, address, players, state, ts, map, mods, maxplayers, title, description, type, width, height, tileset, author) 
             VALUES (:name, :addr, :players, :state, :time, :map, :mods, :maxplayers, :title, :description, :type, :width, :height, :tileset, :author)');
@@ -56,34 +121,38 @@
         $insert->bindValue(':time', time(), PDO::PARAM_INT);
         $insert->bindValue(':map', $_REQUEST['map'], PDO::PARAM_STR);
         $insert->bindValue(':mods', $_REQUEST['mods'], PDO::PARAM_STR);
-        $insert->bindValue(':maxplayers', $_REQUEST['maxplayers'], PDO::PARAM_INT);
-        $insert->bindValue(':title', $_REQUEST['title'], PDO::PARAM_STR);
-        $insert->bindValue(':description', $_REQUEST['description'], PDO::PARAM_STR);
-        $insert->bindValue(':type', $_REQUEST['type'], PDO::PARAM_STR);
-        $insert->bindValue(':width', $_REQUEST['width'], PDO::PARAM_INT);
-        $insert->bindValue(':height', $_REQUEST['height'], PDO::PARAM_INT);
-        $insert->bindValue(':tileset', $_REQUEST['tileset'], PDO::PARAM_STR);
-        $insert->bindValue(':author', $_REQUEST['author'], PDO::PARAM_STR);
+        $insert->bindValue(':maxplayers', $maxplayers, PDO::PARAM_INT);
+        $insert->bindValue(':title', $title, PDO::PARAM_STR);
+        $insert->bindValue(':description', $description, PDO::PARAM_STR);
+        $insert->bindValue(':type', $type, PDO::PARAM_STR);
+        $insert->bindValue(':width', $width, PDO::PARAM_INT);
+        $insert->bindValue(':height', $height, PDO::PARAM_INT);
+        $insert->bindValue(':tileset', $tileset, PDO::PARAM_STR);
+        $insert->bindValue(':author', $author, PDO::PARAM_STR);
         $insert->execute();
         
-        $delete = $db->prepare("DELETE FROM players WHERE server_address = '" . $addr . "'");
-        $delete->execute();
-        
-        $i = 0;
-        while ($i < $players)
+        // title is set, client is supported
+        if (isset( $_REQUEST['title'] ))
         {
-			$playerName = $_REQUEST['playerName'.(string)$i];
-			$playerFaction = $_REQUEST['playerFaction'.(string)$i];
-			$playerTeam = $_REQUEST['playerTeam'.(string)$i];
-			$insert = $db->prepare('INSERT INTO players
-            (server_address, name, faction, team) 
-            VALUES (:server_address, :name, :faction, :team)');
-            $insert->bindValue(':server_address', $addr, PDO::PARAM_STR);
-            $insert->bindValue(':name', $playerName, PDO::PARAM_STR);
-            $insert->bindValue(':faction', $playerFaction, PDO::PARAM_STR);
-            $insert->bindValue(':team', $playerTeam, PDO::PARAM_INT);
-            $insert->execute();
-            $i++;
+			$delete = $db->prepare("DELETE FROM players WHERE server_address = '" . $addr . "'");
+			$delete->execute();
+        
+			$i = 0;
+			while ($i < $players)
+			{
+				$playerName = $_REQUEST['playerName'.(string)$i];
+				$playerFaction = $_REQUEST['playerFaction'.(string)$i];
+				$playerTeam = $_REQUEST['playerTeam'.(string)$i];
+				$insert = $db->prepare('INSERT INTO players
+				(server_address, name, faction, team) 
+				VALUES (:server_address, :name, :faction, :team)');
+				$insert->bindValue(':server_address', $addr, PDO::PARAM_STR);
+				$insert->bindValue(':name', $playerName, PDO::PARAM_STR);
+				$insert->bindValue(':faction', $playerFaction, PDO::PARAM_STR);
+				$insert->bindValue(':team', $playerTeam, PDO::PARAM_INT);
+				$insert->execute();
+				$i++;
+			}
 		}
 		
         if (isset( $_REQUEST['new']))
